Fan used with automatic and manual mode
Small default :
IAQ is going negative on TTN when the value goes over 255. We should send it as luminance/luminosity to be able to have 2 bytes (we do not see the 2 digits precision). We could also put in on 1 byte and modify the recuperation function to to *5.

Humidity is not sended

Motor speed is not sended. We could do it (but maybe only each 1 or 2 minutes)